requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
PyJWT==1.7.1
sqlparse==0.5.0
tzdata==2024.1
python-docx==1.1.2
langdetect
PyPDF2==3.0.1
openai==1.30.1
pillow==10.3.0
Spire.Doc
python-pptx==0.6.23
pdfplumber==0.11.0
pandas==2.2.2